requirements.txt
sentence-transformers
weaviate-client<4.0.0
python-dotenv
pandas
